Refactor beta_banner.js
