+++ scripts/src/beta-banner.js
@@ -1,19 +1,17 @@
+import create_dismiss_button from './modules/beta_banner/create_dismiss_button';
+import extract_usage_policy from './modules/beta_banner/extract_usage_policy';
+
 document.addEventListener("DOMContentLoaded", () => {
+    const accept_cookies_button = document.querySelector("#accept_optional_cookies");
     const beta_banner = document.querySelector(".beta-banner");
+    const cookies_usage_policy = extract_usage_policy(document.cookie);
 
-    if(beta_banner) {
-        const body = document.querySelector("body");
-        const beta_banner_options = document.querySelector(".beta-banner__options");
-        const button = document.createElement("button");
-        const domain = window.location.hostname;
-
-        button.setAttribute("class", "beta-banner__button");
-        button.innerText = "Dismiss this message";
-        beta_banner_options.appendChild(button);
-
-        button.addEventListener("click", () => {
-            body.removeChild(beta_banner);
-            document.cookie = `beta_banner_dismissed=true; domain=${domain}; path=/; Secure`;
+    if(accept_cookies_button) {
+        accept_cookies_button.addEventListener("click", () => {
+            create_dismiss_button(beta_banner);
         })
+    }
+    else if(beta_banner && cookies_usage_policy) {
+        create_dismiss_button(beta_banner);
     }
 });
