+++ scripts/src/modules/beta_banner/extract_usage_policy.js
@@ -1,0 +1,11 @@
+export default function extract_usage_policy(cookies) {
+    if(cookies) {
+        const cookies_array = decodeURIComponent(cookies).split(";").map(index => index.trim());
+        const cookies_policy = cookies_array.filter(index => index.slice(0, 14) === "cookies_policy").pop();
+        const cookies_policy_object = JSON.parse(cookies_policy.slice(cookies_policy.indexOf("=") + 1));
+
+        return cookies_policy_object["usage"];
+    }
+    
+    return false;
+}
