+++ templates/includes/beta-banner.html
@@ -7,7 +7,7 @@
         {{ settings.core.SiteSettings.beta_banner_text|richtext }}
     </div>
     <div class="beta-banner__options">
-        <a class="beta-banner__link" href="{{ settings.core.SiteSettings.beta_banner_link }}">
+        <a class="beta-banner__link" href="{{ settings.core.SiteSettings.beta_banner_link }}" rel="noopener noreferrer" target="_blank">
             {{ settings.core.SiteSettings.beta_banner_link_text }}
         </a>
     </div>
